Refactored setup
