+++ setup.py
@@ -3,5 +3,5 @@ from distutils.core import setup
 setup(name='Pelican representative image plugin',
       version='0.1',
       url='https://github.com/linevich/representative_image',
-      packages=['representative_image',],
+      py_modules = ('representative_image',),
 )
