+++ test_representative_image.py
@@ -13,7 +13,7 @@ TEST_CUSTOM_IMAGE_URL = 'https://testimage.com/custom.jpg'
 
 
 from pelican.contents import Article
-from .representative_image import representative_image
+import representative_image
 
 class TestRepresentativeImage(unittest.TestCase):
